Design to be modular application design
